(1) Ported about half of recmem.cpp (2) Added temporary dumping ground for random constants.
--HG--
extra : convert_revision : svn%3A849a83f0-50c2-4045-bdd1-8b1f698b4b18/trunk%4019

+++ jsoar/src/org/jsoar/kernel/symbols/Identifier.java
@@ -29,8 +29,8 @@ public class Identifier extends Symbol
     public boolean allow_bottom_up_chunks;
     
     public boolean could_be_a_link_from_below;
-    public short level;
-    public short promotion_level;
+    public int level;
+    public int promotion_level;
     public int link_count;
     // TODO: dl_cons unknown_level
     public final ListHead<Slot> slots = new ListHead<Slot>(); // dll of slots for this identifier
